LayoutTests/fast/dom/DeviceMotion/optional-event-properties.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>
<script>
description("Tests the optional properties of DeviceMotionEvent. Each property should be null if not set, or set to null or undefined.");

function ObjectThrowingException() {};
ObjectThrowingException.prototype.valueOf = function() { throw new Error('valueOf threw exception'); }
ObjectThrowingException.prototype.__defineGetter__("x", function() { throw new Error('x getter exception'); });
ObjectThrowingException.prototype.__defineGetter__("alpha", function() { throw new Error('alpha getter exception'); });
var objectThrowingException = new ObjectThrowingException();

function testException(expression, expectedException)
{
    shouldThrow(expression, '(function() { return "' + expectedException + '"; })();');
}

var event;

evalAndLog("event = document.createEvent('DeviceMotionEvent')");
shouldBeTrue("event.acceleration == null");
shouldBeTrue("event.accelerationIncludingGravity == null");
shouldBeTrue("event.rotationRate == null");
shouldBeTrue("event.interval == null");
</script>
</body>
</html>
