LayoutTests/fast/dom/DeviceMotion/optional-event-properties-expected.txt

Tests the optional properties of DeviceMotionEvent. Each property should be null if not set, or set to null or undefined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


event = document.createEvent('DeviceMotionEvent')
PASS event.acceleration == null is true
PASS event.accelerationIncludingGravity == null is true
PASS event.rotationRate == null is true
PASS event.interval == null is true
PASS successfullyParsed is true

TEST COMPLETE
